\name{DeRezende.Ferreira-package}
\alias{DeRezende.Ferreira-package}
\alias{DeRezende.Ferreira}
\docType{package}
\title{Zero Coupon Yield Curve Modelling}
\description{
Modeling the zero coupon yield curve using the dynamic De Rezende and Ferreira (2011) <doi:10.1002/for.1256>
five factor model with variable or fixed decaying parameters. For explanatory purposes, the package also includes
various short datasets of interest rates for the BRICS countries.
}
\details{
\tabular{ll}{
Package: \tab DeRezende.Ferreira\cr
Type: \tab Package\cr
Version: \tab 4.1\cr
Date: \tab 2019-04-26\cr
Depends: \tab R (>= 3.5.0), xts, stats\cr
License: \tab GPL (>= 2)\cr
Encoding: \tab UTF-8\cr
LazyData: \tab true\cr
RoxygenNote: \tab 6.1.1\cr

}

}
\author{
Author: Oleksandr Castello [aut, cre]\cr
        Marina Resta [ctb, cre]

Maintainer: Oleksandr Castello <alexander-castello@libero.it>\cr
}
\references{
De Rezende R.B., Ferreira M.S., “Modeling and Forecasting the Brazilian Term Structure of Interest Rates by an Extended Nelson-Siegel Class of Models: A Quantile Autoregression Approach” (2008).

De Rezende R.B., Ferreira M.S. (2011), “Modeling and Forecasting the Yield Curve by an Extended Nelson-Siegel Class of Models: A Quantile Autoregression Approach”, \emph{Journal of Forecasting  , J. Forecast. n. 32}, \bold{p. 111–123} (2013).

De Rezende R.B., “Giving Flexibility to The Nelson-Siegel Class of Term Structure Models”, \emph{Revista Brasileira de Financas Vol. 9, N. 1 (2011)}, \bold{p. 27–49}.

Caldeira. J. F., Moura G. V., Portugal M.S., “Efficient Yield Curve Estimation and Forecasting in Brazil”, \emph{Revista EconomiA (Brasilia) (January/April 2009)} ,v.11, n.1, \bold{p.27–51}.

Nelson C. R., Siegel A.F., “Parsimonious Modeling of Yield Curves”, \emph{The Journal of Business (1987)}, \bold{60}, 473-489.

Diebold F.X., Li C.,” Forecasting the term structure of government bond yields”, \emph{Journal of Econometrics n. 130}, \bold{337-364} (2005).

Diebold, F.X., Ji, L. and Li, C. , A Three-Factor Yield Curve Model: Non-Affine Structure, Systematic Risk Sources, and Generalized Duration, in L.R. Klein (ed.), \emph{Long-Run Growth and Short-Run Stabilization: Essays in Memory of Albert Ando}. Cheltenham, U.K.: Edward Elgar, \bold{p. 240-274} (2006).

Guirreri S.S., "Modelling and estimation of the yield curve", \emph{Package "YieldCurve" - February 19, 2015}, \bold{CRAN}.
}

