\name{ZC_China}
\alias{ZC_China}
\docType{data}
\title{Zero-Coupon interest rates}
\description{
ZC Government bonds with maturities (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 15, 20, 30) expressed in years and with  business day frequency, source: Thomson Reuters Datastream. The range date is from 2018-01-02 to 2018-01-15.
}
\usage{data(ZC_China)}
\format{
A data frame (txt file) with 14 daily interest rates at different maturities and 10 observed days.
}

