\name{ZC_Russia}
\alias{ZC_Russia}
\docType{data}
\title{Zero-Coupon interest rates}
\description{
ZC Government bonds with maturities (0.25, 0.5, 0.75, 1,2,3,5,7,10,15,20,30) expressed in years and with  business day frequency, source: the Central Bank of the Russian Federation web site. The range date is from 2018-01-03 to 2018-01-17.
}
\usage{data(ZC_Russia)}
\format{
A data frame (txt file) with 12 daily interest rates at different maturities and 10 observed days.
}
