\name{ZC_SouthAfrica}
\alias{ZC_SouthAfrica}
\docType{data}
\title{Zero-Coupon interest rates}
\description{
ZC Government bonds with maturities (0.25, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 15, 20, 25, 30) expressed in years and with  business day frequency, source: Thomson Reuters Datastream. The range date is from 2018-01-01 to 2018-01-12.
}
\usage{data(ZC_SouthAfrica)}
\format{
A data frame (txt file) with 16 daily interest rates at different maturities and 10 observed days.
}

